requirements.txt
Jinja2==2.8
MarkupSafe==0.23
PyTrie==0.2
PyYAML==3.11
Pygments==2.0.2
RPi.GPIO==0.5.11
Twisted==15.4.0
autobahn==0.10.9
cffi==1.2.1
characteristic==14.3.0
click==5.1
crossbar==0.11.1
cryptography==1.0.2
idna==2.0
-e git+git@mameebox-github:/mameebox/mameebox.git@7f98786c1077f1dbcf46d5be0a88d95d999101f4#egg=mameebox-origin/master
mistune==0.7.1
netaddr==0.7.18
path.py==8.1.2
pyOpenSSL==0.15.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyserial==2.7
requests==2.8.0
service-identity==14.0.0
shutilwhich==1.1.0
six==1.10.0
treq==15.0.0
txaio==2.0.3
zope.interface==4.1.3
